requirements.txt
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
cycler==0.11.0
Django==3.2.9
djangorestframework==3.12.4
easyocr==1.4.1
h11==0.12.0
idna==3.3
imageio==2.10.3
kiwisolver==1.3.2
matplotlib==3.4.3
networkx==2.6.3
numpy==1.21.4
opencv-python-headless==4.5.4.58
outcome==1.1.0
Pillow==8.2.0
pycparser==2.20
pyOpenSSL==21.0.0
pyparsing==3.0.4
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.1.1
PyYAML==6.0
requests==2.26.0
scikit-image==0.18.3
scipy==1.7.2
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3
sqlparse==0.4.2
tifffile==2021.11.2
torch==1.10.0
torchvision==0.11.1
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==3.10.0.2
urllib3==1.26.7
wsproto==1.0.0
